refactor(types): improved file structure to types
