+++ src/types/entities/Coordinator.ts
@@ -1,0 +1,6 @@
+export type Coordinator = {
+  id: number
+  name: string
+  courseName: string
+  status: string
+}
